+++ src/Listener/RequestListener.php
@@ -89,8 +89,8 @@ class RequestListener
             throw new InternalErrorException('No site available for the current request with uri '.htmlspecialchars($request->getUri(), ENT_QUOTES));
         }
 
-        if ($site->getLocale() && $site->getLocale() !== $request->get('_locale')) {
-            throw new PageNotFoundException(sprintf('Invalid locale - site.locale=%s - request._locale=%s', $site->getLocale(), $request->get('_locale')));
+        if ($site->getSiteLocale() && $site->getSiteLocale() !== $request->get('_locale')) {
+            throw new PageNotFoundException(sprintf('Invalid locale - site.locale=%s - request._locale=%s', $site->getSiteLocale(), $request->get('_locale')));
         }
 
         try {