+++ src/Site/HostSiteSelector.php
@@ -39,8 +39,8 @@ class HostSiteSelector extends BaseSiteSelector
             }
         }
 
-        if ($this->site && $this->site->getLocale()) {
-            $event->getRequest()->attributes->set('_locale', $this->site->getLocale());
+        if ($this->site && $this->site->getSiteLocale()) {
+            $event->getRequest()->attributes->set('_locale', $this->site->getSiteLocale());
         }
     }
 }