+++ tests/Site/BaseLocaleSiteSelectorTest.php
@@ -54,14 +54,14 @@ abstract class BaseLocaleSiteSelectorTest extends TestCase
         $sites[0]->setRelativePath('/fr');
         $sites[0]->setHost('localhost');
         $sites[0]->setIsDefault(true);
-        $sites[0]->setLocale('fr');
+        $sites[0]->setSiteLocale('fr');
 
         $sites[1] = new Site();
         $sites[1]->setEnabled(true);
         $sites[1]->setRelativePath('/en');
         $sites[1]->setHost('localhost');
         $sites[1]->setIsDefault(false);
-        $sites[1]->setLocale('en');
+        $sites[1]->setSiteLocale('en');
 
         return $sites;
     }